acl-coding

+++ dataplane/src/acl/dp_acl.h
@@ -8,10 +8,6 @@
 #include "mbuf.h"
 
 
-
-
-
-
 #define DP_ACL_RULELIST_NAME "DP_ACL_RULELIST"
 
 
@@ -29,6 +25,6 @@ extern uint32_t DP_Acl_Rule_Init();
 extern uint32_t DP_Acl_Load_Rule(rule_list_t *rule_list,rule_set_t* ruleset, hs_node_t* node);
 extern bool firewall_pass_rule(mbuf_t* p);
 extern uint32_t DP_Acl_Rule_Clean(rule_set_t* ruleset, hs_node_t* node);
-extern uint32_t DP_Acl_Lookup(mbuf_t *mb);
+extern uint8_t DP_Acl_Lookup(mbuf_t *mb);
 
 #endif
